pyproject.toml
[project]
name = "mon-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==4.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "drf-yasg>=1.21.14",
    "nltk>=3.9.2",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "scikit-learn>=1.8.0",
]
